schemas adicionados, substituindo tipos incosistentes nos models

+++ models/cliente.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, ForeignKey
+from sqlalchemy import Column, Integer, Numeric, ForeignKey
 from sqlalchemy.orm import relationship
 from .base import Base
 
@@ -6,7 +6,7 @@ class Cliente(Base):
     __tablename__ = 'clientes'
     cliente_id = Column(Integer, primary_key=True, autoincrement=True)
     usuario_id = Column(Integer, ForeignKey('usuarios.usuario_id'))
-    saldo = Column(Float, default=0.0)
+    saldo = Column(Numeric(10, 2), default=0.0)
     usuario = relationship('Usuario', back_populates='cliente')
     selos = relationship('SelosCliente', back_populates='cliente', uselist=False)
     favoritos = relationship('Favorito', back_populates='cliente')
